Sending-Message/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
twilio==7.16.0
urllib3==1.26.13
